Arel::Nodes::DeleteStatement
