+++ lib/arel/extensions.rb
@@ -53,9 +53,9 @@ require 'arel/extensions/infer'
 require 'arel/extensions/set_to_default'
 require 'arel/extensions/update_statement'
 require 'arel/extensions/current_of_expression'
+require 'arel/extensions/delete_statement'
 
 require 'arel/extensions/functions'
-require 'arel/extensions/nodes'
 require 'arel/extensions/visitors'
 
 module Arel
