--- lib/arel/extensions/nodes.rb
@@ -1,18 +1,0 @@
-# rubocop:disable Metrics/ParameterLists
-# rubocop:disable Naming/UncommunicativeMethodParamName
-# rubocop:disable Metrics/ModuleLength
-
-module Arel
-  module Nodes
-    # https://www.postgresql.org/docs/9.5/sql-insert.html
-    Arel::Nodes::DeleteStatement.class_eval do
-      attr_accessor :using
-      attr_accessor :with
-      attr_accessor :returning
-    end
-  end
-end
-
-# rubocop:enable Metrics/ParameterLists
-# rubocop:enable Naming/UncommunicativeMethodParamName
-# rubocop:enable Metrics/ModuleLength
